server/requirements.txt
fastapi
uvicorn
motor
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pytest
pytest-asyncio


